Add auth middleware to interface with Azure AD B2C
For moderators to be able to login with their own external accounts
(e.g., Google, Facebook, Twitter, etc.), we are using Azure AD B2C.

We also need to make sure moderators accounts are manually approved
before they can use the platform, which is what this middleware does.

A user flow has been set up in Azure AD B2C to call the
/pendingRegistrations endpoint just before a new user is registered
through AD. We then create a pending registration record in our DB and
present the 'block page' to the user informing them that their account
is now pending approval. We also send an email to the 'approver' contact
to ask them to verify the account, using a unique token, which we store
hashed in the DB so nobody knows it except the email recipient.

When the approver verifies the account, they call the
/pendingRegistrations/approve/:token endpoint, which we then use to
create the actual user in Azure AD B2C and send them a notification
email that their account has been approved. We also delete the pending
registration request at this point.

I haven't figured out a way to test this using localhost URLs yet, so
I'm proxing my localhost to a real URL.

Note: I was originally trying to do this as an Azure lambda function,
but we need persistent storage to store the pending registration
requests, and using Azure databases is not free, so seemed like an
unneccessary cost.

Note: another method I tried was to create a user with
'accountEnabled=false' on registration, and setting it to true after
manual verification but this won't work because accountEnabled only
works for local accounts and not externally federated accounts.

Note: another method I tried was to create a custom attribute on the
User object after registration called 'PendingApproval', and adding
another user-flow for login to return the 'block page' if that attribute
is true, but turns out AD doesn't support block pages for login (only
allows it for registration).