Configure NextAuth in moderator-dashboard for Azure B2C
This is more complicated than it should have been, but Azure B2C only
supports the `openid` and `offline_access` scopes, and when using a
social provider, will not provide the user `email` as a claim no matter
what configuration we do in the User Flow in Azure.

The only way to get Azure B2C to return the user email as a claim (it
seems) is to create a custom Policy (XML), but this requires additional
maintenance and manual updates when properties change, which isn't
really feasible for this project.

So, to get the user email, we will make a request to the Graph API as
the application itself (with higher privileges), get the email address,
and add it to the NextAuth profile/session.