Don't allow direct access to any page other than homepage
We want to reduce the risk of abusers seeing e.g., a link in their
device's history that could expose what a survivor has been doing on the
device. Therefore, we want to make sure that any link that is visited
directly (and is not the homepage) does not work, i.e., all pages must
be rendered through client-transitions beginning at the homepage.

This logic utilises Next.js's client transitions which fetch the data for page
transitions via a special _next/data URL.

This does mean that if e.g., a page is refreshed, it will no longer
work, which is the trade-off for this type of privacy feature.

When a page is visited directly, we will go to a 'panic' URL that is not
obvious, currently named `/l.html`. This uses
`document.location.replace` to replace the history entry, so no trace of
the site is added to the browser history at all when such a link is
followed.

Note: I tried to use Next.js middlewares for this, but that does not
return the _next/data URL path so we cannot determine if a client
transition is happening or not.

Note: we cannot use <meta> refresh links to do the final redirect
because these leave a browser history record, which would leave a trace
of the site.