Cleanup port env variables
In production, all containers share the same network, so we do not need
to explicitly forward ports to the host -- they are accessed through the
Docker network using the container names as hostnames.

Also, default all microservices to expose on port 3000 in prod, making
it easier in the gateway config.