Add credentials provider for E2E tests
There does not seem to be an agreed-upon method of testing auth with
NextAuth.js, especially when we're using an OAuth provider.

Some common recommendations are to use an actual account e.g., with
Microsoft but that would require having valid env variables for login
and maintaining that account which does not seem feasible.

Having a credential in test-only seems like a decent trade-off.