Don't rate limit SSR requests
We were previously rate limiting all non-mod/admin API requests,
including those initiated internally through SSR. But this means that on
pages like the forum, by the time you reach the final board listing of
posts, when you click on the post within 2 seconds, you would get rate
limited and it would fail.

We need to filter out SSR requests from the rate limiting logic, and
using the origin IP address is not reliable because the Docker container
of the frontend's IP address is not static (e.g., it's not just
`127.0.0.1`).

So instead, we use @njford's idea: a secret key in a custom header
showing HAProxy that the request originates internally for SSR purposes.